requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.13.2
cssselect2==0.2.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.1
django-adminlte3==0.1.2
django-codemirror2==0.2
django-cors-headers==3.2.0
django-cte==1.1.4
django-database-view==0.1.3
django-easy-pdf==0.2.0.dev1
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-import-export==2.0
django-jet==1.0.8
django-js-asset==1.2.2
django-model-mutations==0.1.1
django-mptt==0.10.0
django-polymorphic==2.1.2
django-report-builder==6.3.2
django-weasyprint==0.5.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
et-xmlfile==1.0.1
Faker==3.0.0
graphene==2.1.8
graphene-django==2.7.1
graphene-django-cud==0.2.1
graphene-django-extras==0.4.8
graphene-file-upload==1.2.2
graphql-core==2.2.1
graphql-relay==2.0.1
gunicorn==20.0.4
html5lib==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mixer==6.1.3
num2words==0.5.10
odfpy==1.4.0
openpyxl==3.0.2
Pillow==7.0.0
promise==2.3
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.13
pylint-plugin-utils==0.6
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
reportlab==3.5.34
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
text-unidecode==1.3
tinycss2==1.0.2
typed-ast==1.4.0
WeasyPrint==51
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
